refactor: extract payment data utility functions to utils.js
