+++ lib/utils.js
@@ -63,3 +63,56 @@ export function formatMonthName(month) {
 export const getUUID = () => {
     return crypto.randomUUID();
 };
+
+export const getTotalAmount = (paymentData) => {
+    const totalIncome = paymentData.reduce((acc, item) => {
+        if (item.amount > 0) {
+            return acc + Math.abs(item.amount);
+        }
+        return acc;
+    }, 0);
+
+    const totalExpense = paymentData.reduce((acc, item) => {
+        if (item.amount < 0) {
+            return acc + Math.abs(item.amount);
+        }
+        return acc;
+    }, 0);
+
+    return { totalIncome, totalExpense };
+};
+
+export const groupByDate = (data) => {
+    const grouped = data.reduce((acc, item) => {
+        const date = item.paidAt.split('T')[0];
+
+        if (!acc[date]) {
+            acc[date] = {
+                date: date,
+                records: [],
+            };
+        }
+
+        acc[date].records.push(item);
+        return acc;
+    }, {});
+
+    return Object.values(grouped).sort(
+        (a, b) => new Date(b.date) - new Date(a.date)
+    );
+};
+
+export const filterPaymentData = ({ data, income, expense }) => {
+    if (income && expense) {
+        return data;
+    }
+
+    if (income) {
+        return data.filter((item) => item.amount > 0);
+    }
+
+    if (expense) {
+        return data.filter((item) => item.amount < 0);
+    }
+    return [];
+};
